+++ src/pages/Auth/Splash.tsx
@@ -1,14 +1,9 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
-import { SERVER_URL } from '../../apis';
-
-const REST_API_KEY = process.env.REACT_APP_REST_API_KEY;
-const REDIRECT_URI = process.env.REACT_APP_REDIRECT_URI;
+import useKakoAuth from '../../hooks/auth/useKakoAuth';
 
 const Splash = () => {
-  const handleKakaoLogin = useCallback(() => {
-    window.location.href = `${SERVER_URL}/oauth2/authorization/kakao?client_id=${REST_API_KEY}&redirect_uri=${REDIRECT_URI}&response_type=code`;
-  }, []);
+  const { handleKakaoLogin } = useKakoAuth();
 
   return (
     <div className='min-h-[90vh]'>
