+++ src/pages/Profile/Profile.tsx
@@ -11,8 +11,8 @@ import { Helmet } from 'react-helmet';
 
 const Profile = () => {
   const { data } = useQuery<UserRequest[]>(['user-requests'], apis.getUserRequests);
-  const { data: UserPostData } = useQuery<UserPostType[]>(['user-posts'], apis.getUserPosts);
-  if (!data || !UserPostData) {
+  const { data: UserPostData, isLoading } = useQuery<UserPostType[]>(['user-posts'], apis.getUserPosts);
+  if (isLoading) {
     return <LoadingSpinner />;
   }
 
@@ -28,13 +28,13 @@ const Profile = () => {
         </Link>
         <div className='bg-[#F4F5F5] h-[8px] w-[100vw] absolute top-[190px]' />
         <div className='flex justify-between w-[100%] mt-[80px] items-center'>
-          <h1 className='font-bold text-start '>내 경기 ({data.length})</h1>
+          <h1 className='font-bold text-start '>내 경기 ({data?.length})</h1>
           <Link to='/profile/userMatchMore'>
             <span className='text-sm'>더보기 +</span>
           </Link>
         </div>
         <UserMatches maxCount={5} />
-        <h1 className='font-bold text-start w-[100%] mt-4'>내가 쓴 게시글 ({UserPostData.length})</h1>
+        <h1 className='font-bold text-start w-[100%] mt-4'>내가 쓴 게시글 ({UserPostData?.length})</h1>
         <UserPost />
       </div>
     </>
