+++ src/pages/Profile/ProfileEdit.tsx
@@ -1,20 +1,14 @@
 import React, { useCallback, useState } from 'react';
-import { useSelector } from 'react-redux';
-import { RootState } from '../../redux/store';
 import { Helmet } from 'react-helmet';
 import UserNicknameChange from '../../components/User/UserNIcknameChange';
-
-interface UpdateUser {
-  nickname: string;
-  deleteImage: string;
-  fileList: FileList;
-}
+import useCurrentUser from '../../hooks/auth/useCurrentUser';
 
 const ProfileEdit = () => {
-  const { nickname } = useSelector((state: RootState) => state.user);
+  const {
+    user: { nickname },
+  } = useCurrentUser();
   const [modalShow, setModalShow] = useState(false);
   const [selected, setSelected] = useState('');
-
   const toggleModal = useCallback((target?: string) => {
     if (target) {
       setSelected(target);
@@ -44,20 +38,6 @@ const ProfileEdit = () => {
             </button>
           </div>
         </div>
-        {/*<div className='flex justify-between items-center mt-8'>*/}
-        {/*  <div>*/}
-        {/*    <h1 className='text-xs'>비밀번호</h1>*/}
-        {/*    <h2 className='text-[#38393C]'>새로운 비밀번호로 변경 가능</h2>*/}
-        {/*  </div>*/}
-        {/*  <div>*/}
-        {/*    <button*/}
-        {/*      className='text-[#38393C] text-[12px] border-[1px] rounded py-1 px-2'*/}
-        {/*      onClick={() => toggleModal('password')}*/}
-        {/*    >*/}
-        {/*      수정*/}
-        {/*    </button>*/}
-        {/*  </div>*/}
-        {/*</div>*/}
       </div>
       <div
         className={`fixed left-0 right-0 bg-white rounded-t-2xl h-[300px] transition-all ${
